Add MSA report generation and SVG visualization features
Implemented functions for generating PDF reports and SVG visualizations for Multiple Sequence Alignment (MSA) results. Refactored existing code to utilize `DocumentManager` for exporting and displaying alignment results, streamlining the process and ensuring consistency across the application.
