+++ composeApp/src/desktopMain/kotlin/com/edu/nwalgo/backend/reportMode/DocumentManager.kt
@@ -1,19 +1,22 @@
 import com.edu.nwalgo.backend.algo.AlignmentResult
-import java.io.File
-import java.io.FileOutputStream
-import java.time.LocalDateTime
-import java.time.format.DateTimeFormatter
+import com.edu.nwalgo.backend.multipleseqaligner.MultipleAlignmentResult
 import com.lowagie.text.Document
 import com.lowagie.text.Font
 import com.lowagie.text.Paragraph
 import com.lowagie.text.pdf.PdfWriter
-import java.awt.BasicStroke
-import java.awt.Color
-import java.awt.Graphics2D
-import java.awt.RenderingHints
+import org.apache.batik.dom.GenericDOMImplementation
+import org.apache.batik.svggen.SVGGraphics2D
+import org.w3c.dom.DOMImplementation
+import java.awt.*
 import java.awt.geom.Path2D
 import java.awt.image.BufferedImage
+import java.io.File
+import java.io.FileOutputStream
+import java.io.FileWriter
+import java.time.LocalDateTime
+import java.time.format.DateTimeFormatter
 import javax.imageio.ImageIO
+import javax.swing.*
 
 /**
  * Manages the generation and export of reports and visualizations related to sequence alignments.
@@ -46,7 +49,7 @@ class DocumentManager {
      *               identity percentage, gap count, and score.
      * @param filename The name of the output PDF file. If not provided, a timestamped filename is generated by default.
      */
-    fun exportAlignmentReportToPDF(result: AlignmentResult, filename: String = "${generateTimestampedFilenamePDF()}") {
+    fun exportAlignmentReportToPDF(result: AlignmentResult, filename: String = generateTimestampedFilenamePDF()) {
         ensureOutputDirectoryExists()
 
         // Export the PDF
@@ -83,7 +86,37 @@ class DocumentManager {
 // Helper function to generate a timestamped filename
     fun generateTimestampedFilenamePDF(): String {
         val formatter = DateTimeFormatter.ofPattern("yyyy-MM-dd_HH-mm-ss")
-        return "report_${LocalDateTime.now().format(formatter)}.pdf"
+        return "MSA_report_${LocalDateTime.now().format(formatter)}.pdf"
+    }
+
+    fun generateMSAReport(result: MultipleAlignmentResult) {
+        ensureOutputDirectoryExists()
+
+        // Export the PDF
+        val filePath = "$RESULTS_FOLDER/${generateTimestampedFilenamePDF()}"
+        val file = File(filePath)
+
+        val document = Document()
+        val writer = PdfWriter.getInstance(document, FileOutputStream(file))
+        document.open()
+
+        val titleFont = Font(Font.HELVETICA, 18f, Font.BOLD)
+        val normalFont = Font(Font.HELVETICA, 12f)
+
+        document.add(Paragraph("Multiple Sequence Alignment Report", titleFont))
+        document.add(Paragraph(" "))
+
+        for (i in result.alignedSequences.indices) {
+            document.add(Paragraph("Aligned Sequence ${i + 1}:\n${result.alignedSequences[i]}", normalFont))
+            document.add(Paragraph(" "))
+        }
+
+        document.add(Paragraph("Identity: %.2f%%".format(result.identity), normalFont))
+        document.add(Paragraph("Gap Count: ${result.gapCount}", normalFont))
+        document.add(Paragraph("Final Score: ${result.score}", normalFont))
+
+        document.close()
+        writer.close()
     }
 
     /**
@@ -115,8 +148,8 @@ class DocumentManager {
 
         // Setting up fonts and rendering
         g.setRenderingHint(RenderingHints.KEY_ANTIALIASING, RenderingHints.VALUE_ANTIALIAS_ON)
-        g.font = java.awt.Font("Arial", java.awt.Font.PLAIN, 18)
-        val letterFont = java.awt.Font("Arial", java.awt.Font.BOLD, 20)
+        g.font = Font("Arial", java.awt.Font.PLAIN, 18)
+        val letterFont = Font("Arial", java.awt.Font.BOLD, 20)
 
         // Filling the background
         g.color = Color(255, 255, 255) // white color
@@ -194,6 +227,7 @@ class DocumentManager {
         return image
 
     }
+
     /**
      * Exports a visual representation of a scoring matrix to a PNG file.
      * The method renders the matrix, sequence labels, and an aligned path (if provided) into an image
@@ -234,6 +268,165 @@ class DocumentManager {
         return "matrix_${LocalDateTime.now().format(formatter)}.png"
     }
 
+    fun generateAlignmentSVG(
+        alignment: List<String>,
+        identity: Double,
+        gapCount: Int,
+        score: Int,
+        outputDir: String = RESULTS_FOLDER,
+        fileName: String = "msa_${generateTimestampedFilenameImage().replace(".png", ".svg")}"
+    ) {
+        ensureOutputDirectoryExists()
+
+        val cellSize = 40
+        val padding = 10
+        val fontSize = 18
+
+        val rows = alignment.size
+        val cols = alignment.firstOrNull()?.length ?: 0
+
+        val width = cols * cellSize + padding * 2
+        val height = rows * cellSize + padding * 2 + 100 // Extra space for stats
+
+        // Prepare output directory
+        val dir = File(outputDir)
+        if (!dir.exists()) {
+            dir.mkdirs()
+        }
+
+        val filePath = File(dir, fileName)
+
+        // Create SVG document
+        val domImpl: DOMImplementation = GenericDOMImplementation.getDOMImplementation()
+        val document: org.w3c.dom.Document = domImpl.createDocument(null, "svg", null)
+        val svgGenerator = SVGGraphics2D(document)
+
+        svgGenerator.setSVGCanvasSize(Dimension(width, height))
+        svgGenerator.font = Font("JetBrains Mono", java.awt.Font.PLAIN, fontSize)
+        val fontMetrics = svgGenerator.fontMetrics
+
+        // Draw alignment matrix
+        for (row in 0 until rows) {
+            for (col in 0 until cols) {
+                val char = alignment[row][col]
+                val x = padding + col * cellSize
+                val y = padding + row * cellSize
+
+                val columnChars = alignment.map { it[col] }
+
+                val color = when {
+                    columnChars.all { it == '-' } -> Color.LIGHT_GRAY
+                    columnChars.distinct().size == 1 -> Color(180, 238, 180)  // Greenish
+                    char == '-' -> Color(220, 220, 220)
+                    else -> Color(255, 160, 160)  // Reddish
+                }
+
+                // Draw background
+                svgGenerator.color = color
+                svgGenerator.fillRect(x, y, cellSize, cellSize)
+
+                // Draw border
+                svgGenerator.color = Color.DARK_GRAY
+                svgGenerator.drawRect(x, y, cellSize, cellSize)
+
+                // Draw character
+                val textWidth = fontMetrics.stringWidth(char.toString())
+                val textHeight = fontMetrics.ascent
+                val textX = x + (cellSize - textWidth) / 2
+                val textY = y + (cellSize + textHeight) / 2 - 4
+
+                svgGenerator.color = Color.BLACK
+                svgGenerator.drawString(char.toString(), textX, textY)
+            }
+        }
+
+        // Draw stats below the alignment
+        val statsY = padding + rows * cellSize + 40
+        svgGenerator.font = Font("SansSerif", java.awt.Font.BOLD, 16)
+        svgGenerator.color = Color.BLACK
+        svgGenerator.drawString("Identity: %.2f%%".format(identity), padding, statsY)
+        svgGenerator.drawString("Gap Count: $gapCount", padding + 200, statsY)
+        svgGenerator.drawString("Score: $score", padding + 400, statsY)
+
+        // Save to file
+        FileWriter(filePath).use { writer ->
+            svgGenerator.stream(writer, true)
+        }
+    }
+
+    fun showMSAVisualization(alignedSequences: List<String>, similarityPercent: Double) {
+        if (alignedSequences.isEmpty()) return
+
+        val rows = alignedSequences.size
+        val cols = alignedSequences.maxOf { it.length }
+        val cellSize = 40
+        val fontSize = 18
+
+        var matches = 0
+        var validCols = 0
+        for (j in 0 until cols) {
+            val column = alignedSequences.mapNotNull { it.getOrNull(j) }.filter { it != '-' }
+            if (column.isNotEmpty()) {
+                validCols++
+                if (column.distinct().size == 1) matches++
+            }
+        }
+
+        val panel = object : JPanel() {
+            override fun getPreferredSize(): Dimension {
+                return Dimension(cols * cellSize + 100, rows * cellSize + 100)
+            }
+
+            override fun paintComponent(g: Graphics) {
+                super.paintComponent(g)
+                val g2 = g as Graphics2D
+                g2.font = Font("JetBrains Mono", java.awt.Font.PLAIN, fontSize)
+
+                for (i in 0 until rows) {
+                    val seq = alignedSequences[i]
+                    for (j in 0 until cols) {
+                        val c = if (j < seq.length) seq[j] else ' '
+                        val isGap = c == '-'
+                        val isMatch = !isGap && alignedSequences.all { it.length > j && it[j] == c }
+
+                        val x = j * cellSize + 50
+                        val y = i * cellSize + 50
+
+                        g2.color = when {
+                            isGap -> Color.LIGHT_GRAY
+                            isMatch -> Color(144, 238, 144)
+                            else -> Color(255, 182, 193)
+                        }
+
+                        g2.fillRect(x, y, cellSize, cellSize)
+                        g2.color = Color.BLACK
+                        g2.drawRect(x, y, cellSize, cellSize)
+                        g2.drawString(c.toString(), x + cellSize / 3, y + cellSize / 2 + 6)
+                    }
+                }
+            }
+        }
+
+        val scrollPane = JScrollPane(panel)
+        scrollPane.preferredSize = Dimension(800, 600)
+
+        // Identity label
+        val identityLabel = JLabel("Identity: %.2f%%".format(similarityPercent))
+        identityLabel.font = Font("SansSerif", java.awt.Font.BOLD, 16)
+        identityLabel.horizontalAlignment = SwingConstants.CENTER
+        identityLabel.border = BorderFactory.createEmptyBorder(10, 0, 10, 0)
+
+        val container = JPanel(BorderLayout())
+        container.add(identityLabel, BorderLayout.NORTH)
+        container.add(scrollPane, BorderLayout.CENTER)
+
+        val frame = JFrame("MSA Visualization")
+        frame.defaultCloseOperation = JFrame.DISPOSE_ON_CLOSE
+        frame.contentPane = container
+        frame.pack()
+        frame.setLocationRelativeTo(null)
+        frame.isVisible = true
+    }
 
 
 }
